LayoutTests/http/tests/loading/resourceLoadStatistics/partitioned-and-unpartitioned-cookie-with-partitioning-timeout-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: Test for Partitioned and Unpartitioned Cookie With User Interaction
main frame - didChangeLocationWithinPageForFrame
CONSOLE MESSAGE: line 63: Step 1
main frame - willPerformClientRedirectToURL: http://localhost:8000/loading/resourceLoadStatistics/resources/set-cookie.php?name=firstPartyCookie&value=value#http://127.0.0.1:8000/loading/resourceLoadStatistics/partitioned-and-unpartitioned-cookie-with-partitioning-timeout.html#step2 
main frame - didFinishDocumentLoadForFrame
main frame - didFinishLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didCancelClientRedirectForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - willPerformClientRedirectToURL: http://127.0.0.1:8000/loading/resourceLoadStatistics/partitioned-and-unpartitioned-cookie-with-partitioning-timeout.html#step2 
main frame - didStartProvisionalLoadForFrame
main frame - didCancelClientRedirectForFrame
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: Test for Partitioned and Unpartitioned Cookie With User Interaction
CONSOLE MESSAGE: line 68: Step 2
main frame - didChangeLocationWithinPageForFrame
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
CONSOLE MESSAGE: line 75: Step 3
main frame - didHandleOnloadEventsForFrame
main frame - didChangeLocationWithinPageForFrame
frame "<!--framePath //<!--frame1-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishDocumentLoadForFrame
CONSOLE MESSAGE: line 82: Step 4
main frame - didChangeLocationWithinPageForFrame
frame "<!--framePath //<!--frame2-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishLoadForFrame
frame "<!--framePath //<!--frame2-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame2-->-->" - didFinishDocumentLoadForFrame
CONSOLE MESSAGE: line 89: Step 5
main frame - didChangeLocationWithinPageForFrame
frame "<!--framePath //<!--frame3-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame2-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame2-->-->" - didFinishLoadForFrame
frame "<!--framePath //<!--frame3-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame3-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame3-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame3-->-->" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
CONSOLE MESSAGE: line 96: Step 6
main frame - didChangeLocationWithinPageForFrame
frame "<!--framePath //<!--frame4-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame4-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame4-->-->" - didFinishDocumentLoadForFrame
CONSOLE MESSAGE: line 103: Step 7
frame "<!--framePath //<!--frame5-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame4-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame4-->-->" - didFinishLoadForFrame
frame "<!--framePath //<!--frame5-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame5-->-->" - didFinishDocumentLoadForFrame
  

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie:

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
Setting partitioned, third-party cookie.


--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
Should only receive partitioned, third-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value

--------
Frame: '<!--framePath //<!--frame3-->-->'
--------
After first user interaction, should only receive un-partitioned cookie.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie: firstPartyCookie=value

--------
Frame: '<!--framePath //<!--frame4-->-->'
--------
Should only receive partitioned, third-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value

--------
Frame: '<!--framePath //<!--frame5-->-->'
--------
After second user interaction, should only receive un-partitioned cookie.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie: firstPartyCookie=value
